Using EthereumClientServer interface
